have to format and add documentation

+++ app/src/main/java/ch/epfl/skysync/screens/flightDetail/FlightDetailsScreen.kt
@@ -7,6 +7,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavHostController
+import ch.epfl.skysync.components.ConfirmFlightDetail
+import ch.epfl.skysync.models.flight.ConfirmedFlight
 import ch.epfl.skysync.navigation.BottomBar
 import ch.epfl.skysync.navigation.Route
 import ch.epfl.skysync.viewmodel.FlightsViewModel
@@ -21,18 +23,28 @@ fun FlightDetailScreen(
   val flight by viewModel.getFlight(flightId).collectAsStateWithLifecycle()
 
   Scaffold(modifier = Modifier.fillMaxSize(), bottomBar = { BottomBar(navController) }) { padding ->
-    FlightDetailUi(
-        backClick = { navController.popBackStack() },
-        deleteClick = {
-          viewModel.deleteFlight(flightId)
-          navController.navigate(Route.HOME)
-        },
-        editClick = { navController.navigate(Route.MODIFY_FLIGHT + "/${flightId}") },
-        confirmClick = { navController.navigate(Route.CONFIRM_FLIGHT + "/${flightId}") },
-        padding = padding,
-        flight = flight,
-        flightId = flightId,
-    )
+      if(flight is ConfirmedFlight){
+          val confirmedFlight = flight as ConfirmedFlight
+          ConfirmFlightDetail(
+              confirmedFlight = confirmedFlight,
+              backClick = { navController.popBackStack() },
+              paddingValues = padding,
+              confirmClick = {}
+          )
+      } else {
+          FlightDetailUi(
+              backClick = { navController.popBackStack() },
+              deleteClick = {
+                  viewModel.deleteFlight(flightId)
+                  navController.navigate(Route.HOME)
+              },
+              editClick = { navController.navigate(Route.MODIFY_FLIGHT + "/${flightId}") },
+              confirmClick = { navController.navigate(Route.CONFIRM_FLIGHT + "/${flightId}") },
+              padding = padding,
+              flight = flight,
+              flightId = flightId,
+          )
+      }
   }
 }
